+++ apps/client/src/routes/my-courses/$courseSlug.tsx
@@ -107,7 +107,7 @@ function LearnPageWrapper() {
     <SidebarProvider
       defaultOpen={!isMobile}
       defaultRightOpen={false}
-      sidebarWidth="22rem"
+      sidebarWidth="24rem"
       headerHeight="4rem"
       className={cn("flex-col", themeClass)}
       style={customStyles}